RavenDB-17519 Performance addnotations.
