+++ src/Raven.Server/Documents/Indexes/Persistence/Lucene/LuceneIndexReadOperation.cs
@@ -64,7 +64,7 @@ namespace Raven.Server.Documents.Indexes.Persistence.Lucene
         {
             try
             {
-                _analyzer = IndexingHelpers.CreateLuceneAnalyzer(index, index.Definition, forQuerying: true);
+                _analyzer = LuceneIndexingHelpers.CreateLuceneAnalyzer(index, index.Definition, forQuerying: true);
             }
             catch (Exception e)
             {
